feat(#37): move compiler and decompiler into corresponding packages
